Fix element selector generator for svg and path (#8)
Selected a different approach to generate selectors for elements, since svg and path tags cannot be searched by the same way as other tag names via xpath for some reason.

Also changed the behaviour for situations, where there is no parent element. Since this should never occur for QApe (it stops at body element, which still has a parent), we will use double slash (//), instead of single slash (/), whenever there is no parent element. This should solve situations, where parent element is not defined for some reason and the element selector generator would think, that this is the full path, but in reality it is not.